# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/inventory": {
      "get": {
        "operationId": "InventoryController_getInventory",
        "summary": "Listar inventário com filtros",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lowStock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inventário retornado com sucesso"
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/inventory/low-stock": {
      "get": {
        "operationId": "InventoryController_getLowStock",
        "summary": "Listar produtos com estoque baixo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Produtos com estoque baixo retornados com sucesso"
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/inventory/movements": {
      "get": {
        "operationId": "InventoryController_getMovements",
        "summary": "Histórico de movimentações de estoque",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "IN",
                "OUT",
                "ADJUSTMENT",
                "RETURN",
                "DAMAGE",
                "TRANSFER"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movimentações retornadas com sucesso"
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "InventoryController_createMovement",
        "summary": "Registrar movimentação de estoque",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMovementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movimentação registrada com sucesso"
          },
          "400": {
            "description": "Dados inválidos ou estoque insuficiente"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/inventory/{productId}": {
      "get": {
        "operationId": "InventoryController_getStock",
        "summary": "Consultar estoque de um produto",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estoque retornado com sucesso"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/inventory/adjust": {
      "post": {
        "operationId": "InventoryController_adjustStock",
        "summary": "Ajustar estoque (gerente/admin)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Estoque ajustado com sucesso"
          },
          "400": {
            "description": "Dados inválidos ou ajuste resultaria em estoque negativo"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "tags": [
          "inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/loyalty/points": {
      "get": {
        "operationId": "LoyaltyController_getAccount",
        "summary": "Obter conta de fidelidade do usuário",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Conta retornada com sucesso"
          }
        },
        "tags": [
          "loyalty"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/loyalty/history": {
      "get": {
        "operationId": "LoyaltyController_getTransactions",
        "summary": "Obter histórico de transações de pontos",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico retornado com sucesso"
          }
        },
        "tags": [
          "loyalty"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/loyalty/redeem": {
      "post": {
        "operationId": "LoyaltyController_redeemPoints",
        "summary": "Resgatar recompensa com pontos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRewardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recompensa resgatada com sucesso"
          },
          "400": {
            "description": "Pontos insuficientes"
          },
          "404": {
            "description": "Recompensa não encontrada"
          }
        },
        "tags": [
          "loyalty"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "summary": "Listar notificações do usuário",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de notificações retornada"
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_getUnreadCount",
        "summary": "Obter contagem de notificações não lidas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Contagem retornada com sucesso"
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "summary": "Marcar notificação como lida",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notificação marcada como lida"
          },
          "404": {
            "description": "Notificação não encontrada"
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_markAllAsRead",
        "summary": "Marcar todas as notificações como lidas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Todas marcadas como lidas"
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/notifications/preferences": {
      "get": {
        "operationId": "NotificationsController_getPreferences",
        "summary": "Obter preferências de notificação",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Preferências retornadas"
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "NotificationsController_updatePreferences",
        "summary": "Atualizar preferências de notificação",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferências atualizadas"
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Supermercado API",
    "description": "API completa para plataforma de supermercado - Venâncio Aires, RS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMovementDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "uuid-do-produto"
          },
          "type": {
            "type": "string",
            "enum": [
              "IN",
              "OUT",
              "ADJUSTMENT",
              "RETURN",
              "DAMAGE",
              "TRANSFER"
            ],
            "example": "IN"
          },
          "quantity": {
            "type": "number",
            "example": 10
          },
          "reason": {
            "type": "string",
            "example": "Reposição de estoque semanal"
          }
        },
        "required": [
          "productId",
          "type",
          "quantity"
        ]
      },
      "AdjustStockDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "uuid-do-produto"
          },
          "quantity": {
            "type": "number",
            "example": -5,
            "description": "Pode ser negativo para redução"
          },
          "reason": {
            "type": "string",
            "example": "Correção após contagem física"
          }
        },
        "required": [
          "productId",
          "quantity",
          "reason"
        ]
      },
...
      "RedeemRewardDto": {
        "type": "object",
        "properties": {
          "rewardId": {
            "type": "string",
            "description": "ID da recompensa",
            "example": "uuid"
          }
        },
        "required": [
          "rewardId"
        ]
      },
      "UpdatePreferencesDto": {
        "type": "object",
        "properties": {
          "orderUpdates": {
            "type": "boolean",
            "description": "Receber atualizações de pedidos"
          },
          "promotions": {
            "type": "boolean",
            "description": "Receber promoções"
          },
          "deliveryUpdates": {
            "type": "boolean",
            "description": "Receber atualizações de entrega"
          },
          "loyaltyUpdates": {
            "type": "boolean",
            "description": "Receber atualizações de fidelidade"
          },
          "pushEnabled": {
            "type": "boolean",
            "description": "Habilitar notificações push"
          },
          "emailEnabled": {
            "type": "boolean",
            "description": "Habilitar notificações por e-mail"
          },
          "smsEnabled": {
            "type": "boolean",
            "description": "Habilitar notificações por SMS"
          }
        }
      },
...
    }
  }
}